more readonly guard coverage

+++ sqlitedict.py
@@ -242,6 +242,9 @@ class SqliteDict(DictClass):
         self.conn.execute(DEL_ITEM, (key,))
 
     def update(self, items=(), **kwds):
+        if self.flag == 'r':
+            raise RuntimeError('Refusing to update read-only SqliteDict')
+
         try:
             items = [(k, encode(v)) for k, v in items.items()]
         except AttributeError:
@@ -256,6 +259,9 @@ class SqliteDict(DictClass):
         return self.iterkeys()
 
     def clear(self):
+        if self.flag == 'r':
+            raise RuntimeError('Refusing to clear read-only SqliteDict')
+
         CLEAR_ALL = 'DELETE FROM %s;' % self.tablename  # avoid VACUUM, as it gives "OperationalError: database schema has changed"
         self.conn.commit()
         self.conn.execute(CLEAR_ALL)
@@ -291,6 +297,9 @@ class SqliteDict(DictClass):
                 pass
 
     def terminate(self):
+        if self.flag == 'r':
+            raise RuntimeError('Refusing to terminate read-only SqliteDict')
+
         """Delete the underlying database file. Use with care."""
         self.close()
